refactor: move `signed_cmp` to `ScalarExt` trait
