+++ crates/proof-of-sql/src/base/scalar/mod.rs
@@ -1,5 +1,5 @@
 /// This module contains the definition of the `Scalar` trait, which is used to represent the scalar field used in Proof of SQL.
-pub mod scalar;
+mod scalar;
 pub use scalar::Scalar;
 mod error;
 pub use error::ScalarConversionError;
@@ -13,3 +13,6 @@ pub(crate) use mont_scalar::MontScalar;
 pub mod test_scalar;
 #[cfg(test)]
 mod test_scalar_test;
+
+mod scalar_ext;
+pub use scalar_ext::ScalarExt;
