+++ crates/proof-of-sql/src/proof_primitive/dory/dory_commitment_test.rs
@@ -1,5 +1,5 @@
 use super::DoryScalar;
-use crate::base::scalar::{Scalar, ScalarConversionError};
+use crate::base::scalar::{Scalar, ScalarConversionError, ScalarExt};
 use core::cmp::Ordering;
 use num_bigint::BigInt;
